@@ -19,7 +19,7 @@ var (
 func init() {
 	cmd.Root.AddCommand(commandDefinition)
 	cmdFlags := commandDefinition.Flags()
-	flags.BoolVarP(cmdFlags, &createEmptySrcDirs, "create-empty-src-dirs", "", createEmptySrcDirs, "Create empty source dirs on destination after copy")
+	flags.BoolVarP(cmdFlags, &createEmptySrcDirs, "create-empty-src-dirs", "", createEmptySrcDirs, "Create empty source dirs on destination after copy", "")
 }
 
 var commandDefinition = &cobra.Command{
@@ -32,40 +32,60 @@ identical on source and destination, testing by size and modification
 time or MD5SUM.  Doesn't delete files from the destination. If you
 want to also delete files from destination, to make it match source,
 use the [sync](/commands/rclone_sync/) command instead.
+
 Note that it is always the contents of the directory that is synced,
 not the directory itself. So when source:path is a directory, it's the
 contents of source:path that are copied, not the directory name and
 contents.
+
 To copy single files, use the [copyto](/commands/rclone_copyto/)
 command instead.
+
 If dest:path doesn't exist, it is created and the source:path contents
 go there.
+
 For example
+
     rclone copy source:sourcepath dest:destpath
+
 Let's say there are two files in sourcepath
+
     sourcepath/one.txt
     sourcepath/two.txt
+
 This copies them to
+
     destpath/one.txt
     destpath/two.txt
+
 Not to
+
     destpath/sourcepath/one.txt
     destpath/sourcepath/two.txt
+
 If you are familiar with |rsync|, rclone always works as if you had
 written a trailing |/| - meaning "copy the contents of this directory".
 This applies to all commands and whether you are talking about the
 source or destination.
+
 See the [--no-traverse](/docs/#no-traverse) option for controlling
 whether rclone lists the destination directory or not.  Supplying this
 option when copying a small number of files into a large destination
 can speed transfers up greatly.
+
 For example, if you have many files in /path/to/src but only a few of
 them change every day, you can copy all the files which have changed
 recently very efficiently like this:
+
     rclone copy --max-age 24h --no-traverse /path/to/src remote:
+
 **Note**: Use the |-P|/|--progress| flag to view real-time transfer statistics.
+
 **Note**: Use the |--dry-run| or the |--interactive|/|-i| flag to test without copying anything.
 `, "|", "`"),
+	Annotations: map[string]string{
+		"groups": "Copy,Filter,Listing,Important",
+	},
 	Run: func(command *cobra.Command, args []string) {
 
 		cmd.CheckArgs(2, 2, command, args)
